+++ org.qiweb/org.qiweb.api/src/main/java/org/qiweb/controllers/package-info.java
@@ -16,4 +16,4 @@
 /**
  * QiWeb Standard Controllers.
  */
-package org.qiweb.api.controllers;
+package org.qiweb.controllers;