+++ org.qiweb/org.qiweb.cli/src/main/java/org/qiweb/cli/DamnSmallDevShell.java
@@ -274,7 +274,7 @@ public final class DamnSmallDevShell
                             + "import org.qiweb.api.outcomes.Outcome;\n\n"
                             + "public class Application {\n\n"
                             + "    public Outcome index() {\n"
-                            + "        return new org.qiweb.api.controllers.Welcome().welcome();\n"
+                            + "        return new org.qiweb.controllers.Welcome().welcome();\n"
                             + "    }\n\n"
                             + "}\n";
         Files.write( new File( ctrlDir, "Application.java" ).toPath(), controller.getBytes( UTF_8 ) );